Fix Project Structure to be only Package.swift
