+++ .github/workflows/test.yml
@@ -10,6 +10,5 @@ jobs:
       - uses: actions/checkout@v4
       - name: Select Xcode version
         run: sudo xcode-select -s "/Applications/Xcode_15.0.1.app/Contents/Developer"
-      - name: Test
-        run: swift test
-        working-directory: CCompiler
+      - name: Test Modules
+        run: make test_modules
